Refactor Growth page to be more modular
