+++ v3_frontend/src/pages/Growth/index.ts
@@ -1,0 +1,6 @@
+export { TotalStatsView } from "./TotalStatsView";
+export { ChronologicalStatsView } from "./ChronologicalStatsView";
+export { GoalsView } from "./GoalsView";
+export { GrowthMenuView } from "./GrowthMenuView";
+export { TagSelectDropdown } from "./TagSelectDropdown";
+export * from "./utils";
